refactor(upload): add src and types folders
