+++ packages/upload/package.json
@@ -4,6 +4,7 @@
   "author": "Evan Sharp <evan.sharp@availity.com>",
   "description": "Availity upload component for uploading files",
   "main": "index.js",
+  "types": "index.d.ts",
   "keywords": [
     "react",
     "availity",
@@ -36,7 +37,6 @@
   },
   "devDependencies": {
     "@availity/upload-core": "^4.1.1",
-    "lodash": "^4.17.21",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-dropzone": "^11.3.2",
