+++ packages/upload/src/FilePickerBtn.js
@@ -1,6 +1,7 @@
 import React, { useRef } from 'react';
 import PropTypes from 'prop-types';
 import { Button, Input } from 'reactstrap';
+
 import FilePicker from './FilePicker';
 
 const FilePickerBtn = ({
