+++ packages/upload/types/FileList.d.ts
@@ -1,0 +1,11 @@
+export interface FileListProps {
+  files?: Array<any>;
+  children?: Function;
+  onRemoveFile?: Function;
+  onPasswordSubmit?: Function;
+  passwordModalZIndex?: number | string;
+}
+
+declare const FileList: React.ComponentType<FileListProps>;
+
+export default FileList;
