+++ packages/upload/types/FilePickerBtn.d.ts
@@ -1,0 +1,15 @@
+export interface FilePickerBtnProps {
+  onClick?: Function;
+  onChange?: Function;
+  multiple?: boolean;
+  name?: string;
+  color?: string;
+  children?: React.ReactType;
+  allowedFileTypes?: Array<string>;
+  maxSize?: number;
+  'data-testid'?: string;
+}
+
+declare const FilePickerBtn: React.ComponentType<FilePickerBtnProps>;
+
+export default FilePickerBtn;
